% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTTQDClient.R
\name{RTTQDClient}
\alias{RTTQDClient}
\title{R6 Class Representing a TickTrader QuotesDownloader Tool}
\description{
A R Client for TickTrader Quote Downlodaer
}
\details{
A R Client to Download Quote History from TickTrader
}
\examples{

## ------------------------------------------------
## Method `RTTQDClient$GetBarsHistory`
## ------------------------------------------------

qdClient <- RTTQDClient$new()
bars <- qdClient$GetBarsHistory("EURUSD", as.Date("2020-01-01"), as.Date("2020-01-02"), "M1", "Bids")

## ------------------------------------------------
## Method `RTTQDClient$GetTickHistory`
## ------------------------------------------------

qdClient <- RTTQDClient$new()
ticks <- qdClient$GetTickHistory("EURUSD", as.Date("2020-01-01"), as.Date("2020-01-02"))
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RTTQDClient$new()}}
\item \href{#method-GetBarsHistory}{\code{RTTQDClient$GetBarsHistory()}}
\item \href{#method-GetTickHistory}{\code{RTTQDClient$GetTickHistory()}}
\item \href{#method-CleanStorage}{\code{RTTQDClient$CleanStorage()}}
\item \href{#method-clone}{\code{RTTQDClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create RTTQDClient object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RTTQDClient$new(
  server = "ttlive.fxopen.com",
  login = "100",
  password = "TTqfdeppmhDR",
  quoteStorage = normalizePath(".//Quotes")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{server}}{Server}

\item{\code{login}}{Login}

\item{\code{password}}{Password}

\item{\code{quoteStorage}}{QuoteStorage Path}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new RTTQDClient object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetBarsHistory"></a>}}
\if{latex}{\out{\hypertarget{method-GetBarsHistory}{}}}
\subsection{Method \code{GetBarsHistory()}}{
Get Bar History
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RTTQDClient$GetBarsHistory(
  symbol,
  startDate = as.Date("2020-01-01"),
  endDate = as.Date("2020-01-02"),
  periodicity = "M1",
  type = "Bids"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{SymbolName.}

\item{\code{startDate}}{StartDate. Date object}

\item{\code{endDate}}{EndDate. Date object}

\item{\code{periodicity}}{Periodicity. One from c("S1", "S10", "M1", "M5", "M15", "M30", "H1", "H4", "D1", "W1", "MN1")}

\item{\code{type}}{Type. One from c("Bids", "Asks", "BidAsk")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table with BarsHistory
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{qdClient <- RTTQDClient$new()
bars <- qdClient$GetBarsHistory("EURUSD", as.Date("2020-01-01"), as.Date("2020-01-02"), "M1", "Bids")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetTickHistory"></a>}}
\if{latex}{\out{\hypertarget{method-GetTickHistory}{}}}
\subsection{Method \code{GetTickHistory()}}{
Get Ticks History
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RTTQDClient$GetTickHistory(
  symbol,
  startDate = Sys.Date() - days(1),
  endDate = Sys.Date()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{SymbolName.}

\item{\code{startDate}}{StartDate. Date object}

\item{\code{endDate}}{EndDate. Date object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table with TicksHistory
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{qdClient <- RTTQDClient$new()
ticks <- qdClient$GetTickHistory("EURUSD", as.Date("2020-01-01"), as.Date("2020-01-02"))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CleanStorage"></a>}}
\if{latex}{\out{\hypertarget{method-CleanStorage}{}}}
\subsection{Method \code{CleanStorage()}}{
Clean local Storage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RTTQDClient$CleanStorage(path = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path.}}{Set if your local}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RTTQDClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
